HSVIforOSPOSGs

Documentation for HSVIforOSPOSGs.

HSVIforOSPOSGs.HSVI

HSVIforOSPOSGs.OSPOSG

HSVIforOSPOSGs.OSPOSG

HSVIforOSPOSGs.Partition

HSVIforOSPOSGs.Recorder

HSVIforOSPOSGs.State

HSVIforOSPOSGs.LB_min

HSVIforOSPOSGs.UB_max

HSVIforOSPOSGs.lipschitz_delta

HSVIforOSPOSGs.solve

HSVIforOSPOSGs.write

HSVIforOSPOSGs.HSVI — Type

HSVI

Heuristic search value iteration solver for OSPOSGs.

- neighborhood: parameter that guarantees Lipschitz continuity and convergence of the algorithm
- presolve_epsilon: precision the solver is trying to achieve during presolve phase
- presolve_time_limit: time limit for the presolve phase in seconds
- optimizer_factory: function that returns optimizer compatible with JuMP.Model

HSVIforOSPOSGs.OSPOSG — Type

OSPOSG

Type for a one-sided partially observable stochastic game. Can be loaded from .osposg files.

HSVIforOSPOSGs.OSPOSG — Method

OSPOSG(path::AbstractString)

Construct OSPOSG from .osposg file at path.

HSVIforOSPOSGs.Partition — Type

Partition

Type for a partition of one-sided partially observable stochastic game.

HSVIforOSPOSGs.Recorder — Type

Recorder

Stores the solver statistics after each iteration.

HSVIforOSPOSGs.State — Type

State

Type for a state of one-sided partially observable stochastic game.

HSVIforOSPOSGs.LB_min — Method

LB_min(osposg::OSPOSG)

Returns the minimal possible value of the game.

HSVIforOSPOSGs.UB_max — Method

UB_max(osposg::OSPOSG)

Returns the maximal possible value of the game.

HSVIforOSPOSGs.lipschitz_delta — Method

lipschitz_delta(osposg::OSPOSG)

Returns the Lipschitz delta of the game.

HSVIforOSPOSGs.solve — Method

solve(osposg::OSPOSG, hsvi::HSVI, epsilon::Float64, time_limit::Float64)

Run hsvi solver on osposg with time limit time_limit aiming for gap of at most epsilon in initial belief.

HSVIforOSPOSGs.write — Method

write(path::AbstractString, osposg::OSPOSG)

Writes game osposg in the .osposg format to path.