Manual

This page shows how to use the HSVIforOSPOSGs.

Getting started

First, a game in .osgposg format must be loaded as OSPOSG:

julia> osposg = OSPOSG("../../games/pursuit-evasion/peg03.osposg")OSPOSG:
  discount = 0.95
  state_count = 143
  partition_count = 21
  player1_action_count = 145
  player2_action_count = 13
  observation_count = 2
  transition_count = 2671
  reward_count = 2671
  minimal_reward = 0.0
  maximal_reward = 100.0
  LB_min = 0.0
  UB_max = 1999.9999999999982
  lipschitz_delta = 999.9999999999991
  initial_partition = 5
  initial_belief = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0]

HSVIforOSPOSGs.OSPOSG — Type

OSPOSG

Type for a one-sided partially observable stochastic game. Can be loaded from .osposg files.

HSVIforOSPOSGs.OSPOSG — Method

OSPOSG(path::AbstractString)
OSPOSG(io::IO)

Construct OSPOSG from .osposg file at path or from IO io.

Then, HSVI algorithm must be instantiated/configured. Here we use HSVI with default parameters:

julia> hsvi = HSVI()HSVI:
  neighborhood = 1.0e-6
  presolve_epsilon = 0.01
  presolve_time_limit = 300.0
  optimizer_factory = #26

For available parameters and their meaning see below:

HSVIforOSPOSGs.HSVI — Type

HSVI

Heuristic search value iteration solver for OSPOSGs.

The constructor takes following keyword arguments:

neighborhood - parameter that guarantees Lipschitz continuity and convergence of the algorithm

presolve_epsilon - precision the solver is trying to achieve during presolve phase

presolve_time_limit - time limit for the presolve phase in seconds

optimizer_factory - function that returns optimizer compatible with JuMP.Model to be used as LP solver

The neigborhood parameter $D$ of HSVI must be within these bounds

\[0 < D < \frac{(1 − \gamma) \varepsilon}{2 \delta}\]

for the algorithm to have convergence guarantees, where $\gamma$ is discount factor, $\varepsilon$ is specified gap and $\delta$ is Lipschitz delta of the game.

After loading the game and configuring the algorithm we can start solving by running solve:

julia> recorder = solve(osposg, hsvi, 1.0, 60.0)┌ Info: OSPOSG:
│   discount = 0.95
│   state_count = 143
│   partition_count = 21
│   player1_action_count = 145
│   player2_action_count = 13
│   observation_count = 2
│   transition_count = 2671
│   reward_count = 2671
│   minimal_reward = 0.0
│   maximal_reward = 100.0
│   LB_min = 0.0
│   UB_max = 1999.9999999999982
│   lipschitz_delta = 999.9999999999991
│   initial_partition = 5
└   initial_belief = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0]
┌ Info: HSVI:
│   neighborhood = 1.0e-6
│   presolve_epsilon = 0.01
│   presolve_time_limit = 300.0
└   optimizer_factory = #26
ERROR: type GitHubActionsLogger has no field stream

HSVIforOSPOSGs.solve — Function

solve(osposg::OSPOSG, hsvi::HSVI, epsilon::Float64, time_limit::Float64)

Run hsvi solver on osposg game with time limit time_limit aiming for gap of at most epsilon in initial belief.

The recorder contains records of the algorithm progress after each iteration and can be used for further examination or presentation of results.

HSVIforOSPOSGs.Recorder — Type

Recorder

Stores the solver statistics after each iteration.

LP Solvers

HSVIforOSPOSGs uses GLPK as a JuMP solver for linear programs by default because it is open-source and easily installable through standalone Julia package. However, note that the LP models being solved in HSVI for OSPOSGs are quite complex and GLPK might get stuck on some of them or report them as being unfeasible (although they should be feasible). Therefore, it is recommended to use more powerful solvers (such as CPLEX, Gurobi, etc.). However, installing them is not as straightforward and often requires 3rd party binary and/or proprietary license. If you wish to use different solver you can do so by installing the JuMP wrapper for the given solver (refer to installation instructions at CPLEX.jl, Gurobi.jl, etc.) and then pass the optimizer factory into HSVI constructor:

using CPLEX

HSVI(optimizer_factory=() -> CPLEX.Optimizer())

Logging level

This package utilizes Julia logging facilities to communicate the progress of the algorithm to the user. By default only info, warning and error messages are displayed. To display more detailed debuging info, use the following before running the algorithm:

using Logging

global_logger(ConsoleLogger(stdout, Logging.Debug))

On the contrary, to disable info logging use:

global_logger(ConsoleLogger(stdout, Logging.Warn))

Examples

Additional examples on how to configure and run the algorithm are provided in the scripts directory.